:recycle: reafactor: move getTranslatedText to `src`
