+++ api/lib/infrastructure/repositories/area-repository.js
@@ -1,7 +1,7 @@
 import { Area } from '../../domain/models/Area.js';
 import { areaDatasource } from '../datasources/learning-content/area-datasource.js';
 import * as competenceRepository from '../../../src/shared/infrastructure/repositories/competence-repository.js';
-import { getTranslatedKey } from '../../domain/services/get-translated-text.js';
+import { getTranslatedKey } from '../../../src/shared/domain/services/get-translated-text.js';
 import _ from 'lodash';
 import { NotFoundError } from '../../domain/errors.js';
